Getting closer to running angular2! :D
Documenting failure of compiling angular2.

I'm not sure what if its the module loading or what but it's far from
compatible with the expressjs setup I had. Also the tutorials have a
tsconfig.json which visual studio 2013 node.js project did not when it
was created.

This is going to be "fun" to clean up after finally getting the angular2
code to run. Files all over the place. At least it's learningful about
how it works!