Can do the multiple test!
Congratulations you beat the test!

Made a reviewService to review the test and it will emit an onreview
event for the question components.

Removed the old code so its only the angular app now.
Next part is re-implement single test. Ideally abstracting the question
& test part into a single controller.

Then for the finale make a gh-pages with a database removal & static
html file as a demo for the blog!