Simple tests work too
Added simpletests.
Added IQuestion interface as shared base for the two types of questions
for the ReviewService.

Backend does not support random between simple & multiple questions but
should be simple to implement in future at least!

Quite a few other backend changes that would be nice to add like a
descriptive tag.  For example the word Te - You(Plural)  it collides
with the word Sinä - You    so the descriptive tag (Plural) would be
nice since currently one has to answer You (Plural) if doing simple.